requirements.txt
# Voice Bot - Server Dependencies

# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# Environment variables
python-dotenv==1.0.0

# Speech-to-Text (Deepgram)
deepgram-sdk>=3.0.0

# Text-to-Speech (gTTS - Free Google TTS)
gTTS==2.5.4

# File operations
aiofiles>=23.0.0

# AI Agent (LangGraph + Groq)
langchain>=0.1.0
langchain-groq>=0.1.0
langgraph>=0.0.20

# For production deployments (optional)
# gunicorn==21.2.0

# Optional: Add other API client libraries as needed:
# openai==1.3.0              # For OpenAI Whisper/GPT/TTS
# anthropic==0.7.0           # For Claude
# google-cloud-speech==2.21.0  # For Google STT
# google-cloud-texttospeech==2.14.1  # For Google TTS
# elevenlabs==0.2.26         # For ElevenLabs TTS

